Update commit on master

+++ src/components/CustomerList.js
@@ -1,12 +1,27 @@
 import React, { useState, useEffect } from 'react';
 import MaterialTable from 'material-table';
-import { Toolbar, Typography } from '@material-ui/core';
-import { Redirect } from 'react-router-dom';
+import { Toolbar, Typography, Button, Tooltip, IconButton } from '@material-ui/core';
+import { Redirect, Link } from 'react-router-dom';
+import CustomerForm from './CustomerForm';
+import DeleteIcon from '@material-ui/icons/Delete';
+import EditIcon from '@material-ui/icons/Edit';
+import AddIcon from '@material-ui/icons/Add';
 
 function CustomerList () {
 
     const [customer, setCustomer] = useState([]);
     const [redirect, setRedirect] = useState([]);
+    const emptyCustomer = {
+        id: 0,
+        firstname: "",
+        lastname: "",
+        streetaddress: "",
+        postcode: "",
+        city: "",
+        phone: "",
+        email: "",
+        link: "https://customerrest.herokuapp.com/api/customers"
+    }
 
     useEffect( () => getCustomers(), []);
 
@@ -16,8 +31,8 @@ function CustomerList () {
         .then(data => setCustomer(data.content))
     }
 
-    const addCustomer = (newCustomer) => {
-        fetch('https://customerrest.herokuapp.com/api/customers', {
+    const addCustomer = (newCustomer, link) => {
+        fetch(link, {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json',
@@ -41,9 +56,11 @@ function CustomerList () {
     }
 
     const deleteCustomer = (link) => {
-        fetch(link, { method: 'DELETE' })
-        .then(res => getCustomers())
-        .catch(err => console.log(err))
+        if (window.confirm("Are you sure you want to delete this customer?")) {
+            fetch(link, { method: 'DELETE' })
+            .then(res => getCustomers())
+            .catch(err => console.log(err))
+        }
     }
 
     const selectCustomer = (event, rowData) => {
@@ -66,8 +83,10 @@ function CustomerList () {
                     Customers
                 </Typography>
             </Toolbar>
+            <CustomerForm formTitle="Add customer" icon={<AddIcon />} handleSubmit={addCustomer} customer={emptyCustomer} />
             <MaterialTable
                 columns={[
+                    { title: "Actions", field: "actions" },
                     { title: "First name", field: "firstname" },
                     { title: "Last name", field: "lastname" },
                     { title: "Street address", field: "streetaddress" },
@@ -76,43 +95,28 @@ function CustomerList () {
                     { title: "Email", field: "email" },
                     { title: "Phone", field: "phone" }
                 ]}
-                data={customer}
-                title="Customers"
-                editable={
-                    {
-                        onRowAdd: newData =>
-                            new Promise( (resolve, reject) => {
-                                setTimeout( () => {
-                                    addCustomer(newData);
-                                    resolve();
-                                }, 1000);
-                        }),
-                        onRowUpdate: (newData, oldData) => 
-                            new Promise( (resolve, reject) => {
-                                setTimeout( () => {
-                                    {
-                                        const index = customer.indexOf(oldData);
-                                        const link = customer[index].links[0].href;
-                                        updateCustomer(newData, link);
-                                        resolve();
-                                    }
-                                    resolve();
-                                }, 1000);
-                        }),
-                        onRowDelete: oldData => 
-                            new Promise( (resolve, reject) => {
-                                setTimeout( () => {
-                                    {
-                                        const index = customer.indexOf(oldData);
-                                        const link = customer[index].links[0].href;
-                                        deleteCustomer(link);
-                                        resolve();
-                                    }
-                                    resolve();
-                                }, 1000);
-                        })
+                data={customer.map((cust) => {
+                    cust.link = cust.links[0].href;
+                    const customerLink = "/customer/id/" + cust.link.split("/").pop();
+                    return {
+                        actions: <div>
+                        <CustomerForm formTitle="Edit" icon={<EditIcon />} handleSubmit={updateCustomer} customer={cust} />
+                        <Tooltip title="Delete">
+                            <IconButton aria-label="delete" onClick={() => deleteCustomer(cust.link)}>
+                                <DeleteIcon />
+                            </IconButton>
+                        </Tooltip>
+                        </div>,
+                        firstname: <Link to={customerLink}>{cust.firstname}</Link>,
+                        lastname: <Link to={customerLink}>{cust.lastname}</Link>,
+                        streetaddress: cust.streetaddress,
+                        postcode: cust.postcode,
+                        city: cust.city,
+                        phone: cust.phone,
+                        email: cust.email
                     }
-                }
+                })}
+                title="Customers"
                 options={
                     { sorting: true,
                      pageSize: 20,
@@ -122,7 +126,7 @@ function CustomerList () {
                         fontWeight: 'bold'
                     } }
                 }
-                onRowClick={(event, rowData) => selectCustomer(event, rowData)}
+                
             />
         </div>
     );
